logistics_supply-bpms-grafana/requirements.txt
# Backend
fastapi==0.68.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic==1.8.2
python-dotenv==0.19.0
psycopg2-binary==2.9.1
redis==3.5.3
requests==2.26.0

# Data Processing
pandas==1.3.3
numpy==1.21.2
scipy==1.7.1

# Machine Learning & Optimization
scikit-learn==0.24.2
tensorflow==2.6.0
ortools==9.2.9972
xgboost==1.4.2
prophet==1.0.1

# OLAP & Database
apache-druid==0.0.1
apache-kylin==2.6.6
clickhouse-driver==0.2.3

# ETL
apache-airflow==2.1.3
prefect==0.15.9
dbt-core==0.21.0
dbt-postgres==0.21.0

# Visualization
dash==2.0.0
plotly==5.3.1
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0

# Infrastructure & Monitoring
prometheus-client==0.11.0
grafana-api==1.0.3
docker==5.0.2
kubernetes==18.20.0
boto3==1.18.44

# Testing
pytest==6.2.5
pytest-cov==2.12.1
black==21.9b0
flake8==3.9.2
mypy==0.910

# Documentation
sphinx==4.2.0
sphinx-rtd-theme==0.5.2

# Development
jupyter==1.0.0
notebook==6.4.4
ipython==7.28.0 